leetcode009修改

+++ src/com/N1njaC/LeetCode009.java
@@ -5,13 +5,13 @@ package com.N1njaC;
  */
 public class LeetCode009 {
 
-    //判断一个int类型的是不是回文数。只要利用反转int数，然后判断一下就行了。其实不用担心溢出情况，因为正反都一样。
+    //判断一个int类型的是不是回文数。只要利用反转int数，然后判断一下就行了。反转之后的数可能会溢出，所以结果还是采用long类型
     public boolean isPalindrome(int x) {
         if (x < 0) {
             return false;
         }
         //12321
-        int result = 0;
+        long result = 0;
         int tmp = x;
         while (tmp != 0) {
             result = result * 10 + tmp % 10;
